Moving APIFeatures class into a separate module
